Eshaq/ChangeTheTypeOfSomeEntities (#127)

+++ src/Datiss.Budget.Entities/DWH/IncomeForcastWs.cs
@@ -20,13 +20,13 @@ namespace Datiss.Budget.Entities.DWH
         
         public int UnitUser { get; set; }
 
-        public int WasteInstallIncome { get; set; }
+        public long WasteInstallIncome { get; set; }
 
-        public int WasteBranchIncome { get; set; }
+        public long WasteBranchIncome { get; set; }
 
-        public int WasteNote3Income { get; set; }
+        public long WasteNote3Income { get; set; }
 
-        public int WsNote11Income { get; set; }
+        public long WsNote11Income { get; set; }
 
         #endregion
