+++ src/Datiss.Budget.Services/Models/IncomeCurrentWsHDTO.cs
@@ -28,23 +28,23 @@ namespace Datiss.Budget.Services.Models
 
         public int ConsumptionUser { get; set; }
 
-        public int Cost { get; set; }
+        public long Cost { get; set; }
 
-        public int Income { get; set; }
+        public long Income { get; set; }
 
-        public int SubscriptionIncome { get; set; }
+        public long SubscriptionIncome { get; set; }
 
-        public int Note3Price { get; set; }
+        public long Note3Price { get; set; }
 
-        public int Note3Income { get; set; }
+        public long Note3Income { get; set; }
 
-        public int SeasonalIncome { get; set; }
+        public long SeasonalIncome { get; set; }
 
-        public int TIncome { get; set; }
+        public long TIncome { get; set; }
 
-        public int Note7Price { get; set; }
+        public long Note7Price { get; set; }
 
-        public int Note7Income { get; set; }
+        public long Note7Income { get; set; }
     }
 
     public class UpdateIncomeCurrentWsHDTO : CreateIncomeCurrentWsHDTO
@@ -80,22 +80,22 @@ namespace Datiss.Budget.Services.Models
 
         public int ConsumptionUser { get; set; }
 
-        public int Cost { get; set; }
+        public long Cost { get; set; }
 
-        public int Income { get; set; }
+        public long Income { get; set; }
 
-        public int SubscriptionIncome { get; set; }
+        public long SubscriptionIncome { get; set; }
 
-        public int Note3Price { get; set; }
+        public long Note3Price { get; set; }
 
-        public int Note3Income { get; set; }
+        public long Note3Income { get; set; }
 
-        public int SeasonalIncome { get; set; }
+        public long SeasonalIncome { get; set; }
 
-        public int TIncome { get; set; }
+        public long TIncome { get; set; }
 
-        public int Note7Price { get; set; }
+        public long Note7Price { get; set; }
 
-        public int Note7Income { get; set; }
+        public long Note7Income { get; set; }
     }
 }
