+++ src/Datiss.Budget.ViewModels/IncomeCurrentWsNHViewModel.cs
@@ -44,15 +44,15 @@ namespace Datiss.Budget.ViewModels
 
         [Required(ErrorMessageResourceType = typeof(ViewModelString), ErrorMessageResourceName = "Required")]
         [Range(0, int.MaxValue, ErrorMessageResourceType = typeof(ViewModelString), ErrorMessageResourceName = "isNumber")]
-        public int Cost { get; set; }
+        public long Cost { get; set; }
 
         [Required(ErrorMessageResourceType = typeof(ViewModelString), ErrorMessageResourceName = "Required")]
         [Range(0, int.MaxValue, ErrorMessageResourceType = typeof(ViewModelString), ErrorMessageResourceName = "isNumber")]
-        public int Income { get; set; }
+        public long Income { get; set; }
 
         [Required(ErrorMessageResourceType = typeof(ViewModelString), ErrorMessageResourceName = "Required")]
         [Range(0, int.MaxValue, ErrorMessageResourceType = typeof(ViewModelString), ErrorMessageResourceName = "isNumber")]
-        public int SubscriptionIncome { get; set; }
+        public long SubscriptionIncome { get; set; }
 
         [Required(ErrorMessageResourceType = typeof(ViewModelString), ErrorMessageResourceName = "Required")]
         [Range(0, int.MaxValue, ErrorMessageResourceType = typeof(ViewModelString), ErrorMessageResourceName = "isNumber")]
@@ -60,19 +60,19 @@ namespace Datiss.Budget.ViewModels
 
         [Required(ErrorMessageResourceType = typeof(ViewModelString), ErrorMessageResourceName = "Required")]
         [Range(0, int.MaxValue, ErrorMessageResourceType = typeof(ViewModelString), ErrorMessageResourceName = "isNumber")]
-        public int SeasonalIncome { get; set; }
+        public long SeasonalIncome { get; set; }
 
         [Required(ErrorMessageResourceType = typeof(ViewModelString), ErrorMessageResourceName = "Required")]
         [Range(0, int.MaxValue, ErrorMessageResourceType = typeof(ViewModelString), ErrorMessageResourceName = "isNumber")]
-        public int Note3Price { get; set; }
+        public long Note3Price { get; set; }
 
         [Required(ErrorMessageResourceType = typeof(ViewModelString), ErrorMessageResourceName = "Required")]
         [Range(0, int.MaxValue, ErrorMessageResourceType = typeof(ViewModelString), ErrorMessageResourceName = "isNumber")]
-        public int Note3Income { get; set; }
+        public long Note3Income { get; set; }
 
         [Required(ErrorMessageResourceType = typeof(ViewModelString), ErrorMessageResourceName = "Required")]
         [Range(0, int.MaxValue, ErrorMessageResourceType = typeof(ViewModelString), ErrorMessageResourceName = "isNumber")]
-        public int TotalIncome { get; set; }
+        public long TotalIncome { get; set; }
 
         public IEnumerable<SelectListItem> UserTypeSource { get; set; }
 
@@ -129,15 +129,15 @@ namespace Datiss.Budget.ViewModels
 
         public string ConsumptionUserDisplay => ConsumptionUser.ToString("N0");
         
-        public int Cost { get; set; }
+        public long Cost { get; set; }
 
         public string CostDisplay => Cost.ToString("N0");
 
-        public int Income { get; set; }
+        public long Income { get; set; }
 
         public string IncomeDisplay => Income.ToString("N0");
 
-        public int SubscriptionIncome { get; set; }
+        public long SubscriptionIncome { get; set; }
 
         public string SubscriptionIncomeDisplay => SubscriptionIncome.ToString("N0");
 
@@ -145,19 +145,19 @@ namespace Datiss.Budget.ViewModels
 
         public string ExcessIncomeDisplay => ExcessIncome.ToString("N0");
 
-        public int SeasonalIncome { get; set; }
+        public long SeasonalIncome { get; set; }
 
         public string SeasonalIncomeDisplay => SeasonalIncome.ToString("N0");
 
-        public int Note3Price { get; set; }
+        public long Note3Price { get; set; }
 
         public string Note3PriceDisplay => Note3Price.ToString("N0");
        
-        public int Note3Income { get; set; }
+        public long Note3Income { get; set; }
 
         public string Note3IncomeDisplay => Note3Income.ToString("N0");
 
-        public int TotalIncome { get; set; }
+        public long TotalIncome { get; set; }
 
         public string TotalIncomeDisplay => TotalIncome.ToString("N0");
     }
